Add. `keepalive` plugin
