+++ .travis.yml
@@ -29,7 +29,7 @@ before_install:
   - source here/bin/activate
 
 install:
-  - luarocks make rockspecs/pegasus-0.9.2-1.rockspec
+  - luarocks make rockspecs/pegasus-scm-0.rockspec
 
 before_script:
   - luarocks show luacheck         > /dev/null 2>&1 || luarocks install luacheck
